requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
frozenlist==1.3.3
fsspec==2023.6.0
gcsfs==2023.6.0
google-api-core==2.11.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
idna==3.4
kafka-python==2.0.2
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
pandas==2.0.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
